Refactoring in Acceptance Tests
