+++ test/acceptance_test/ue_functions.py
@@ -1,30 +1,9 @@
-import lte_pb2
-import rrc_pb2
+import zmq
 
-def attach_ue(ue_id, socket):
-    req = lte_pb2.ASN1()
-    req.rrc.attach_req.id = ue_id 
-    socket.send(req.SerializeToString())
-
-    message = socket.recv()
-    resp = lte_pb2.ASN1()
-    resp.ParseFromString(message)
+def create_ue(port):
+    context = zmq.Context()
+    ue = context.socket(zmq.DEALER)
+    ue.bind("tcp://*:%s" % port)
+    ue.RCVTIMEO = 1000
     
-    return resp.rrc.attach_resp
-
-def detach_ue(ue_id, socket):
-    req = lte_pb2.ASN1()
-    req.rrc.detach_req.id = ue_id 
-    socket.send(req.SerializeToString())
-
-    message = socket.recv()
-    resp = lte_pb2.ASN1()
-    resp.ParseFromString(message)
-    
-    return resp.rrc.detach_resp
-
-def uplink_ue(ue_id, socket, data):
-    req = lte_pb2.ASN1()
-    req.rrc.uplink.id = ue_id
-    req.rrc.uplink.data = data
-    socket.send(req.SerializeToString())
+    return ue, port, port
